+++ packages/block-editor/README.md
@@ -527,6 +527,10 @@ _Returns_
 
 -   `Object`: Typography block support derived CSS classes & styles.
 
+### GlobalStylesContext
+
+Undocumented declaration.
+
 ### InnerBlocks
 
 _Related_
@@ -788,6 +792,18 @@ _Returns_
 
 -   `any`: value
 
+### useGlobalStylesOutput
+
+Undocumented declaration.
+
+### useGlobalStylesReset
+
+Undocumented declaration.
+
+### useGlobalStylesSetting
+
+Undocumented declaration.
+
 ### useInnerBlocksProps
 
 This hook is used to lightly mark an element as an inner blocks wrapper
@@ -827,6 +843,10 @@ _Parameters_
 _Returns_
 
 -   `any`: Returns the value defined for the setting.
+
+### useStyle
+
+Undocumented declaration.
 
 ### Warning
