+++ packages/block-editor/src/components/index.js
@@ -161,3 +161,4 @@ export { default as __experimentalInspectorPopoverHeader } from './inspector-pop
 
 export { default as BlockEditorProvider } from './provider';
 export { default as useSetting } from './use-setting';
+export * from './global-styles'; // TODO: Should this instead be in @wordpress/global-styles?
