+++ packages/edit-site/src/components/global-styles/custom-css.js
@@ -8,11 +8,7 @@ import {
 	PanelBody,
 } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
-
-/**
- * Internal dependencies
- */
-import { useStyle } from './hooks';
+import { useStyle } from '@wordpress/block-editor';
 
 function CustomCSSControl() {
 	const [ customCSS, setCustomCSS ] = useStyle( 'css' );
