+++ packages/edit-site/src/components/global-styles/index.js
@@ -1,3 +1,1 @@
 export { default as GlobalStylesUI } from './ui';
-export { useGlobalStylesReset, useStyle, useSetting } from './hooks';
-export { useGlobalStylesOutput } from './use-global-styles-output';
