+++ packages/edit-site/src/components/global-styles/screen-background-color.js
@@ -7,7 +7,11 @@ import classnames from 'classnames';
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { __experimentalColorGradientControl as ColorGradientControl } from '@wordpress/block-editor';
+import {
+	__experimentalColorGradientControl as ColorGradientControl,
+	useGlobalStylesSetting as useSetting,
+	useStyle,
+} from '@wordpress/block-editor';
 
 /**
  * Internal dependencies
@@ -17,8 +21,6 @@ import {
 	getSupportedGlobalStylesPanels,
 	useColorsPerOrigin,
 	useGradientsPerOrigin,
-	useSetting,
-	useStyle,
 } from './hooks';
 
 function ScreenBackgroundColor( { name, variationPath = '' } ) {
